+++ spug_web/src/pages/deploy/request/Ext1Form.js
@@ -249,6 +249,7 @@ export default observer(function () {
         )}
       </Form>
       {visible && <HostSelector
+        title="可选主机列表"
         host_ids={host_ids}
         app_host_ids={app_host_ids}
         onCancel={() => setVisible(false)}
