+++ spug_web/src/pages/deploy/request/Table.js
@@ -3,27 +3,21 @@
  * Copyright (c) <spug.dev@gmail.com>
  * Released under the AGPL-3.0 License.
  */
-import React from 'react';
+import React, { useState } from 'react';
 import { observer } from 'mobx-react';
 import { BranchesOutlined, BuildOutlined, TagOutlined, PlusOutlined, TagsOutlined } from '@ant-design/icons';
-import { Radio, Modal, Popover, Tag, Popconfirm, Tooltip, message } from 'antd';
+import { Radio, Modal, Popover, Tag, Tooltip, Button, Space, message } from 'antd';
 import { http, hasPermission } from 'libs';
 import { Action, AuthButton, TableCard } from 'components';
+import HostSelector from './HostSelector';
 import S from './index.module.less';
 import store from './store';
 import moment from 'moment';
-
-function DeployConfirm() {
-  return (
-    <div>
-      <div>确认发布方式</div>
-      <div style={{color: '#999', fontSize: 12}}>补偿：仅发布上次发布失败的主机。</div>
-      <div style={{color: '#999', fontSize: 12}}>全量：再次发布所有主机。</div>
-    </div>
-  )
-}
+import lds from 'lodash';
 
 function ComTable() {
+  const [request, setRequest] = useState()
+
   const columns = [{
     title: '申请标题',
     className: S.min180,
@@ -144,6 +138,14 @@ function ComTable() {
               <Action.Button auth="deploy.request.do" onClick={() => store.rollback(info)}>回滚</Action.Button>
             )}
           </Action>;
+        case '4':
+          return <Action>
+            <Action.Button auth="deploy.request.do" onClick={() => store.readConsole(info)}>查看</Action.Button>
+            <DoAction info={info}/>
+            {info.visible_rollback && (
+              <Action.Button auth="deploy.request.do" onClick={() => store.rollback(info)}>回滚</Action.Button>
+            )}
+          </Action>;
         case '-1':
           return <Action>
             <Action.Button auth="deploy.request.edit" onClick={() => store.showForm(info)}>编辑</Action.Button>
@@ -171,17 +173,23 @@ function ComTable() {
   }];
 
   function DoAction(props) {
-    const {host_ids, fail_host_ids} = props.info;
+    const {deploy_status} = props.info;
     return (
-      <Popconfirm
-        title={<DeployConfirm/>}
-        okText="全量"
-        cancelText="补偿"
-        cancelButtonProps={{disabled: [0, host_ids.length].includes(fail_host_ids.length)}}
-        onConfirm={e => handleDeploy(e, props.info, 'all')}
-        onCancel={e => handleDeploy(e, props.info, 'fail')}>
+      <Popover trigger="click" zIndex={2} title="确认发布方式" content={(
+        <div>
+          <div style={{color: '#999', fontSize: 12}}>全量：发布所有主机（包含已成功的）。</div>
+          <div style={{color: '#999', fontSize: 12}}>补偿：仅发布上次发布失败的主机。</div>
+          <div style={{color: '#999', fontSize: 12}}>灰度：选择指定主机发布。</div>
+          <Space style={{width: '100%', justifyContent: 'flex-end', marginTop: 16}}>
+            <Button size="small" disabled={!lds.findKey(deploy_status, x => x !== '2')}
+                    onClick={() => handleDeploy(props.info, 'fail')}>补偿</Button>
+            <Button ghost size="small" type="primary" onClick={() => setRequest(props.info)}>灰度</Button>
+            <Button size="small" type="primary" onClick={() => handleDeploy(props.info, 'all')}>全量</Button>
+          </Space>
+        </div>
+      )}>
         <Action.Button auth="deploy.request.do">发布</Action.Button>
-      </Popconfirm>
+      </Popover>
     )
   }
 
@@ -199,43 +207,58 @@ function ComTable() {
     })
   }
 
-  function handleDeploy(e, info, mode) {
+  function handleDeploy(info, mode) {
+    if (request && mode.length === 0) {
+      return message.error('请选择灰度发布的主机')
+    }
     info.mode = mode
-    store.showConsole(info);
+    store.showConsole(info)
+    if (request) setRequest()
   }
 
   return (
-    <TableCard
-      tKey="dr"
-      rowKey={row => row.key || row.id}
-      title="申请列表"
-      columns={columns}
-      scroll={{x: 1500}}
-      tableLayout="auto"
-      loading={store.isFetching}
-      dataSource={store.dataSource}
-      onReload={store.fetchRecords}
-      actions={[
-        <AuthButton
-          auth="deploy.request.add"
-          type="primary"
-          icon={<PlusOutlined/>}
-          onClick={() => store.addVisible = true}>新建申请</AuthButton>,
-        <Radio.Group value={store.f_status} onChange={e => store.f_status = e.target.value}>
-          <Radio.Button value="all">全部({store.counter['all'] || 0})</Radio.Button>
-          <Radio.Button value="0">待审核({store.counter['0'] || 0})</Radio.Button>
-          <Radio.Button value="1">待发布({store.counter['1'] || 0})</Radio.Button>
-          <Radio.Button value="3">发布成功({store.counter['3'] || 0})</Radio.Button>
-          <Radio.Button value="-3">发布异常({store.counter['-3'] || 0})</Radio.Button>
-          <Radio.Button value="99">其他({store.counter['99'] || 0})</Radio.Button>
-        </Radio.Group>
-      ]}
-      pagination={{
-        showSizeChanger: true,
-        showLessItems: true,
-        showTotal: total => `共 ${total} 条`,
-        pageSizeOptions: ['10', '20', '50', '100']
-      }}/>
+    <React.Fragment>
+      <TableCard
+        tKey="dr"
+        rowKey={row => row.key || row.id}
+        title="申请列表"
+        columns={columns}
+        scroll={{x: 1500}}
+        tableLayout="auto"
+        loading={store.isFetching}
+        dataSource={store.dataSource}
+        onReload={store.fetchRecords}
+        actions={[
+          <AuthButton
+            auth="deploy.request.add"
+            type="primary"
+            icon={<PlusOutlined/>}
+            onClick={() => store.addVisible = true}>新建申请</AuthButton>,
+          <Radio.Group value={store.f_status} onChange={e => store.f_status = e.target.value}>
+            <Radio.Button value="all">全部({store.counter['all'] || 0})</Radio.Button>
+            <Radio.Button value="0">待审核({store.counter['0'] || 0})</Radio.Button>
+            <Radio.Button value="1">待发布({store.counter['1'] || 0})</Radio.Button>
+            <Radio.Button value="3">发布成功({store.counter['3'] || 0})</Radio.Button>
+            <Radio.Button value="-3">发布异常({store.counter['-3'] || 0})</Radio.Button>
+            <Radio.Button value="99">其他({store.counter['99'] || 0})</Radio.Button>
+          </Radio.Group>
+        ]}
+        pagination={{
+          showSizeChanger: true,
+          showLessItems: true,
+          showTotal: total => `共 ${total} 条`,
+          pageSizeOptions: ['10', '20', '50', '100']
+        }}/>
+
+      {request ? (
+        <HostSelector
+          title="选择灰度发布的主机"
+          app_host_ids={request.host_ids}
+          onCancel={() => setRequest()}
+          deploy_status={request.deploy_status}
+          onOk={ids => handleDeploy(request, ids)}/>
+      ) : null}
+    </React.Fragment>
   )
 }
